requirements.txt
Django==3.2.10
django-bulma==0.8.3
flake8-django==1.1.5
black==22.12.0
djangorestframework
markdown
django-filter
django-extensions